Move str to its own file
